add ( investment_blue)gold ) : add SelectedItems
